test colors

+++ inc/variables.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   variables.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aolabarr <aolabarr@student.42.fr>          +#+  +:+       +#+        */
+/*   By: beiglesi <beiglesi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/01 14:30:53 by beiglesi          #+#    #+#             */
-/*   Updated: 2025/04/12 15:27:05 by aolabarr         ###   ########.fr       */
+/*   Updated: 2025/04/17 11:17:11 by beiglesi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@
 
 typedef struct s_color
 {
-	double		red;
-	double		green;
-	double		blue;
+	float		red;
+	float		green;
+	float		blue;
 }			t_color;
 
 typedef struct s_ambient
